refactor(browser): encapsulate playwright/axe laziness in the package
Move the playwright-core and axe-core value imports behind dynamic
import() inside @react-doctor/browser (load-playwright.ts + runAxe), so
the barrel has no heavy static deps and consumers can import it
statically. This dissolves the duplicated loadBrowser/isModuleNotFoundError
dance in both the CLI and MCP (the missing-playwright hint now lives in
one place) and lets parseViewport + MAX_VIEWPORT_PX live once in the
browser package, reused by the CLI's Commander wrapper and the MCP tool.

Bundle boundary preserved: axe-core splits into lazy chunks and
playwright-core stays an external dynamic import — neither lands in the
main cli.js startup path.
