+++ packages/browser/src/index.ts
@@ -1,4 +1,5 @@
 export { BrowserSession } from "./session.js";
 export { connectToBrowser } from "./connect.js";
 export type { BrowserConnection } from "./connect.js";
+export { parseViewport } from "./parse-viewport.js";
 export type * from "./types.js";
