+++ packages/mcp/src/utils/with-session.ts
@@ -1,5 +1,4 @@
-import type { BrowserSession, Viewport } from "@react-doctor/browser";
-import { loadBrowser } from "./load-browser.js";
+import { BrowserSession, type Viewport } from "@react-doctor/browser";
 
 export interface BrowserToolConnection {
   cdp?: string;
@@ -10,12 +9,13 @@ export interface BrowserToolConnection {
 // Attach a fresh session per tool call, act, then disconnect. The page lives in
 // the browser (attached over CDP), so each call is cheap and the page persists
 // across calls — the same persistent model the CLI's `browser` commands use.
+// playwright-core loads lazily inside the session, so a missing install surfaces
+// the browser package's own actionable hint (caught by `runTool`).
 export const withSession = async <ResultType>(
   connection: BrowserToolConnection,
   useSession: (session: BrowserSession) => Promise<ResultType>,
 ): Promise<ResultType> => {
-  const { BrowserSession: Session } = await loadBrowser();
-  const session = await Session.attach({
+  const session = await BrowserSession.attach({
     cdpEndpoint: connection.cdp,
     launch: connection.noLaunch === true ? false : undefined,
   });
